move transaction-detail constants to transaction file
