+++ src/constants/index.js
@@ -14,4 +14,3 @@ export * from './patron-agency';
 export * from './transactions';
 export * from './api';
 export * from './search';
-export * from './transaction-detail';
